Adds super simple analyzers
Adds analyzers that check

* String literals calling Markup
* Command line option argument templates

Had to make a couple of the parsers public in the main Spectre project to gain access to them, and I'm relying on catching exceptions to show the error but it seems to work.

VERY rough. Basic scenarios with string literals are practically hard coded, but it has unit tests and a good framework to build on if someone is a bit more clever with roslyn than me